requirements.txt
asgiref==3.5.0
azure-storage-blob==12.11.0
certifi==2021.5.30
charset-normalizer==2.0.0
cloudpickle==2.0.0
cycler==0.11.0
bayesian-optimization==1.2.0
Django==3.2.3
hyperopt==0.2.5
joblib==0.17.0
lightgbm==2.3.1
matplotlib==3.5.1
mssql-django==1.1.1
numpy==1.19.0
openpyxl==3.0.5
pandas==1.1.0
plotly==5.4.0
pyswarms==1.2.0
pytz==2022.1
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
tqdm==4.64.0
typing_extensions==4.1.1
whitenoise==6.1.0
xgboost==0.82